+++ pyhpo/set.py
@@ -1,6 +1,8 @@
+import warnings
+
 from pyhpo.ontology import Ontology
 from pyhpo.term import HPOTerm
-import warnings
+from pyhpo.matrix import Matrix
 
 
 class HPOSet(set):
@@ -95,11 +97,14 @@ class HPOSet(set):
         for term in self:
             if term.is_obsolete:
                 try:
-                    replaced = Ontology[HPOTerm.id_from_string(term.replaced_by)]
+                    replaced = Ontology[
+                        HPOTerm.id_from_string(term.replaced_by)
+                    ]
                     ids.add(replaced)
                 except AttributeError:
                     warnings.warn(
-                        'The term {} is obsolete and has no replacement.'.format(term),
+                        'The term {} is obsolete and has no replacement.'
+                            .format(term),
                         UserWarning)
 
             else:
@@ -311,7 +316,7 @@ class HPOSet(set):
             for term_b in self._list[i+1:]:
                 yield (term_a, term_b)
 
-    def similarity(self, other, kind='omim', method=None):
+    def similarity(self, other, kind=None, method=None, combine='funSimAvg'):
         """
         Calculates the similarity to another HPOSet
         According to Robinson et al, American Journal of Human Genetics, (2008)
@@ -322,30 +327,27 @@ class HPOSet(set):
         other: HPOSet
             Another HPOSet to measure the similarity to
 
-        kind: str, default ``omim``
+        kind: str, default ``None``
             Which kind of information content should be calculated.
             Options are ['omim', 'orpha', 'decipher', 'gene']
+            See :func:`pyhpo.HPOTerm.similarity_score` for options
 
-        method: string, default ``resnik``
+        method: string, default ``None``
             The method to use to calculate the similarity.
+            See :func:`pyhpo.HPOTerm.similarity_score` for options
+
+            Additional options:
+
+            * **equal** - Calculates exact matches between both sets
+
+        combine: string, default ``funSimAvg``
+            The method to combine similarity measures.
 
             Available options:
 
-            * **resnik** - Resnik P, Proceedings of the 14th IJCAI, (1995)
-            * **lin** - Lin D, Proceedings of the 15th ICML, (1998)
-            * **jc** - Jiang J, Conrath D, ROCLING X, (1997)
-              Implementation according to R source code
-            * **jc2** - Jiang J, Conrath D, ROCLING X, (1997)
-              Implementation according to paper from R ``hposim`` library
-              Deng Y, et. al., PLoS One, (2015)
-            * **rel** - Relevance measure - Schlicker A, et.al.,
-              BMC Bioinformatics, (2006)
-            * **ic** - Information coefficient - Li B, et. al., arXiv, (2010)
-            * **graphic** - Graph based Information coefficient -
-              Deng Y, et. al., PLoS One, (2015)
-            * **dist** - Distance between terms
-            * **equal** - Calculates exact matches between both sets
-
+            * **funSimAvg** - Schlicker A, BMC Bioinformatics, (2006)
+            * **funSimMax** - Schlicker A, BMC Bioinformatics, (2006)
+            * **BMA** - Deng Y, et. al., PLoS One, (2015)
 
         Returns
         -------
@@ -356,9 +358,37 @@ class HPOSet(set):
         if method == 'equal':
             return self._equality_score(other)
 
-        score1 = HPOSet._sim_score(self, other, kind, method)
-        score2 = HPOSet._sim_score(other, self, kind, method)
-        return (score1 + score2)/2
+        score_matrix = HPOSet._sim_score(self, other, kind, method)
+
+        row_maxes = [
+            max([v for v in row])
+            for row in score_matrix.rows
+        ]
+
+        col_maxes = [
+            max([v for v in col])
+            for col in score_matrix.columns
+        ]
+
+        if combine == 'funSimAvg':
+            return (
+                sum(row_maxes)/len(row_maxes) +
+                sum(col_maxes)/len(col_maxes)
+            )/2
+
+        if combine == 'funSimMax':
+            return max([
+                sum(row_maxes)/len(row_maxes),
+                sum(col_maxes)/len(col_maxes)
+            ])
+
+        if combine == 'BMA':
+            return (
+                (sum(row_maxes) + sum(col_maxes)) /
+                (len(row_maxes) + len(col_maxes))
+            )
+
+        raise RuntimeError('Invalid combine method specified')
 
     def _equality_score(self, other):
         """
@@ -382,6 +412,8 @@ class HPOSet(set):
             The similarity score to the other HPOSet
 
         """
+        if not len(self) or not len(other):
+            return 0
 
         matches = 0
         for term1 in self:
@@ -391,9 +423,9 @@ class HPOSet(set):
         return matches / max([len(self), len(other)])
 
     @staticmethod
-    def _sim_score(set1, set2, kind, method=None):
+    def _sim_score(set1, set2, kind=None, method=None):
         """
-        Calculates one-way similarity from one HPOSet to another HPOSet
+        Calculates similarity matrix between HPOSets
 
         .. warning::
 
@@ -409,10 +441,11 @@ class HPOSet(set):
 
         kind: str
             Which kind of information content should be calculated.
-            Options are ['omim', 'orpha', 'decipher', 'gene']
+            See :function:`pyhpo.HPOTerm.similarity_score` for options
 
-        method: string, default ``resnik``
+        method: string
             The method to use to calculate the similarity.
+            See :function:`pyhpo.HPOTerm.similarity_score` for options
 
         Returns
         -------
@@ -422,16 +455,16 @@ class HPOSet(set):
         """
 
         if not len(set1) or not len(set2):
-            return 0
+            return Matrix(0, 0)
 
         scores = []
         for set1_term in set1:
-            scores.append(0)
             for set2_term in set2:
-                score = set1_term.similarity_score(set2_term, kind, method)
-                if score > scores[-1]:
-                    scores[-1] = score
-        return sum(scores)/len(scores)
+                scores.append(
+                    set1_term.similarity_score(set2_term, kind, method)
+                )
+
+        return Matrix(len(set1), len(set2), scores)
 
     @classmethod
     def from_queries(cls, queries):